+++ packages/node-utils/src/index.ts
@@ -13,4 +13,4 @@ export {
 export { checkFileExists } from './checkFileExists';
 export { checkSocks5Proxy } from './checkSocks5Proxy';
 export { validateJsonSchema } from './validateJsonSchema';
-export { findProcessFromIncomingPort, type ProcessInfo } from './findProcessFromIncomingPort';
+export { findProcessFromIncomingPort, type ProcessInfo } from './process';
